backend/requirements.txt
fastapi
uvicorn
pydantic
openai
huggingface_hub
langchain-huggingface
langchain-chroma
chromadb
pysqlite3-binary
sentence-transformers